refactor(contract): move integration tests to src/test/
Verified: precommit-verify ✓ (1fd9688de1e94b65)
